Touch tracking works, ABVector[] is passed to ABKeyboard for processing
