+++ AccessBraille/AccessBraille/ABKeyboard.m
@@ -8,15 +8,21 @@
 
 #import "ABKeyboard.h"
 #import "ABActivateKeyboardGestureRecognizer.h"
+#import "ABTypes.h"
 
 @implementation ABKeyboard {
     
 }
 
-- (id)init {
+- (id)initWithDelegate:(id)delegate {
     self = [super init];
     if (self) {
         // Init code
+        _delegate = delegate;
+        ABActivateKeyboardGestureRecognizer *activate = [[ABActivateKeyboardGestureRecognizer alloc] initWithTarget:self action:@selector(ABKeyboardRecognized:)];
+        [activate setTouchDelegate:self];
+        [((UIViewController *)_delegate).view addGestureRecognizer:activate];
+
     }
     return self;
 }
@@ -31,12 +37,33 @@
 }
 
 /**
- * Gets touch colums from GR
+ * Creates a view overlay for recognizing braille type
  */
-- (void)touchColumnsWithInfo:(NSDictionary *)info {
-    NSLog(@"%@", [info[ABGestureInfoStatus] boolValue] ? @"Valid" : @"Invalid");
+- (UIView *)viewWithTouchesFromABVectorArray:(ABVector[])vectors {
+    
+    // Set up overlay view
+    UIView *interface = [[UIView alloc] initWithFrame:[UIScreen mainScreen].bounds];
+    [interface setBackgroundColor:[UIColor grayColor]];
+    [interface setAlpha:0.4];
+    
+    // Test view
+    
+    
+    
+    return interface;
 }
 
+/**
+ * Gets touch colums from GR
+ */
+- (void)touchColumns:(ABVector[])vectors withInfo:(NSDictionary *)info{
+    NSLog(@"%@", info);
+    NSLog(@"%@", ABVectorPrintable(vectors));
+}
+
+/**
+ * recieving end to activation gestures
+ */
 - (void)ABKeyboardRecognized:(ABActivateKeyboardGestureRecognizer *)reg {
     switch (reg.activateDirection) {
         case ABGestureDirectionUP:
