+++ AccessBraille/AccessBraille/NSArray+ObjectSubsets.h
@@ -14,7 +14,7 @@
 + (NSArray *)arrayFromArray:(NSArray *)array passingTest:(BOOL (^)(id obj1))compare;
 
 /* Adds arrays together and keeps them sorted **ONLY UITOUCHES** */
-+ (NSArray *)addToArray:(NSArray *)arrayOrg from:(NSArray *)arrayAdd inView:(UIView *)view;
++ (NSArray *)addToArrayCopiesfrom:(NSArray *)arrayAdd withReferanceToView:(UIView *)view intoArray:(NSArray *)parentArray;
 
 /* Makes a nice one line statement */
 - (NSString *)oneLineNSStringOfArray;
